Mod: `LaneInfo` moved to new file - it also use for `AudioMixer.ts`
